Corrected errors and added examples

+++ oters_lang/src/interpret/allocator.rs
@@ -3,8 +3,8 @@ use std::collections::{BinaryHeap, HashSet};
 
 #[derive(Debug)]
 pub(super) struct Allocator {
-    free_locs: BinaryHeap<Reverse<u64>>,
-    max_loc: u64,
+    free_locs: BinaryHeap<Reverse<u32>>,
+    max_loc: u32,
 }
 
 impl Allocator {
@@ -15,7 +15,7 @@ impl Allocator {
         }
     }
 
-    pub(super) fn alloc(&mut self) -> u64 {
+    pub(super) fn alloc(&mut self) -> u32 {
         match self.free_locs.pop() {
             None => {
                 self.max_loc += 1;
@@ -25,7 +25,7 @@ impl Allocator {
         }
     }
 
-    pub(super) fn dealloc_set(&mut self, locs: HashSet<u64>) {
+    pub(super) fn dealloc_set(&mut self, locs: HashSet<u32>) {
         for loc in locs {
             self.free_locs.push(Reverse(loc))
         }
